fix(Core/Misc): Allow uint32 for NextPage to match DB page_text

+++ src/server/game/Globals/ObjectMgr.h
@@ -56,7 +56,7 @@ struct PlayerLevelInfo;
 struct PageText
 {
     std::string Text;
-    uint16 NextPage;
+    uint32 NextPage;
 };
 
 /// Key for storing temp summon data in TempSummonDataContainer
